feat: reorganize project structure and update README for Podcast Mode
- Created core files (main.py, models.py, inference.py, app.py) into a new `src/` directory with an __init__.py file to form a proper package.
- Updated README.md:
  - Added updated information on Podcast Mode with details on speaker configuration.
  - Included images for UI screenshots.
  - Clarified installation and usage instructions.
- Ensured overall modularity and maintainability.
