+++ run.bat
@@ -1,0 +1,3 @@
+@echo off
+python -m src.main
+pause
