+++ run.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+# run.sh: Launch the Llasa TTS application using the src package
+
+# Ensure the script is run from the project root directory
+python -m src.main
