+++ src/__init__.py
@@ -1,0 +1,1 @@
+# Intentionally left empty
